refactor: 부모컴포넌트에 함수 받아서 자식으로 props로 전달
